Implement CGPatch with tests
Relocated cgpatch and reverted removal of cgc

More feedback included

removed virtual call filter

used static variable for initialization

Applied clang-format

Accumulation of minor bug fixes

1. Fixes a bug that occurred when de-serializing MetaCG graphs into that had FileProperties Metadata attached
2. Fixes a bug in which the check whether to include a function in the graph was omitted if the node was a callee
3. Fixes includes to be compatible with newer LLVM versions

Note: This disables the Alias-Analysis Tests

[MetaCG][NFC] Add second-level READMEs

Adds readmes for the different "sub projects" so that not all of the
text needs to be present in the top-level file.
The idea is that we have a better and clearer structure in the repo and
that a first-time visitor is not overloaded with unnecessary details.

renamed cgpatch wrappers

Fixes and added README

fixed tests

fixed tests

Added preproc & cmake variable for MPI and cmake variable to enablÃe CGPatch

cgpatch now supports llvm18

Reverted changes to MetaCollector

Improved SymbolRetriever

Rule of 5

reformat